+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes in GAP.jl
 
+## Version 0.17.3-DEV (released 2026-MM-DD)
+
+- Honor `Overrides.toml` UUID/name overrides for `GAP_pkg_*` artifacts
+  when locating GAP package directories
+
 ## Version 0.17.2 (released 2026-06-21)
 
 - Update the `alnuth` GAP package to 4.0.2